COMMERCE-7304 Move all providers to the right packages
