+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/taglib/clay/data/set/PlacedCommerceOrderItemDataSetActionProvider.java
@@ -12,23 +12,25 @@
  * details.
  */
 
-package com.liferay.commerce.order.content.web.internal.frontend;
+package com.liferay.commerce.order.content.web.internal.frontend.taglib.clay.data.set;
 
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.order.content.web.internal.frontend.constants.CommerceOrderDataSetConstants;
 import com.liferay.commerce.order.content.web.internal.frontend.util.CommerceOrderClayTableUtil;
-import com.liferay.commerce.order.content.web.internal.model.Order;
+import com.liferay.commerce.order.content.web.internal.model.OrderItem;
+import com.liferay.commerce.service.CommerceOrderService;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetActionProvider;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemBuilder;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.WebKeys;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -37,14 +39,14 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Luca Pellizzon
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.provider.key=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PENDING_ORDERS,
+	property = "clay.data.provider.key=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PLACED_ORDER_ITEMS,
 	service = ClayDataSetActionProvider.class
 )
-public class PendingCommerceOrderDataSetActionProvider
+public class PlacedCommerceOrderItemDataSetActionProvider
 	implements ClayDataSetActionProvider {
 
 	@Override
@@ -52,29 +54,37 @@ public class PendingCommerceOrderDataSetActionProvider
 			HttpServletRequest httpServletRequest, long groupId, Object model)
 		throws PortalException {
 
-		List<DropdownItem> dropdownItems = new ArrayList<>();
+		OrderItem orderItem = (OrderItem)model;
 
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		Order order = (Order)model;
-
-		if (_modelResourcePermission.contains(
-				themeDisplay.getPermissionChecker(), order.getOrderId(),
-				ActionKeys.VIEW)) {
-
-			dropdownItems.add(
-				DropdownItemBuilder.setHref(
-					CommerceOrderClayTableUtil.getEditOrderURL(
-						order.getOrderId(), httpServletRequest)
-				).setLabel(
-					LanguageUtil.get(httpServletRequest, "view")
-				).build());
+		if (orderItem.getParentOrderItemId() > 0) {
+			return Collections.emptyList();
 		}
 
-		return dropdownItems;
+		CommerceOrder commerceOrder = _commerceOrderService.getCommerceOrder(
+			orderItem.getOrderId());
+
+		return DropdownItemListBuilder.add(
+			() -> _modelResourcePermission.contains(
+				PermissionThreadLocal.getPermissionChecker(), commerceOrder,
+				ActionKeys.VIEW),
+			dropdownItem -> {
+				ThemeDisplay themeDisplay =
+					(ThemeDisplay)httpServletRequest.getAttribute(
+						WebKeys.THEME_DISPLAY);
+
+				dropdownItem.setHref(
+					CommerceOrderClayTableUtil.getViewShipmentURL(
+						orderItem.getOrderItemId(), themeDisplay));
+
+				dropdownItem.setLabel(
+					LanguageUtil.get(httpServletRequest, "shipments"));
+				dropdownItem.setTarget("modal");
+			}
+		).build();
 	}
+
+	@Reference
+	private CommerceOrderService _commerceOrderService;
 
 	@Reference(
 		target = "(model.class.name=com.liferay.commerce.model.CommerceOrder)"
